+++ frontend/src/components/Course/Teachers/TeacherList.tsx
@@ -1,10 +1,10 @@
 import { Col, Form, ListGroup, OverlayTrigger, Row, Tooltip } from "react-bootstrap"
 import { DashCircle, DashCircleFill, PlusCircleFill } from 'react-bootstrap-icons';
-import { useAddOwnerMutation, useAddTutorMutation, useGetOwnersQuery, useGetTutorsQuery, useRemoveOwnerMutation, useRemoveTutorMutation } from "../features/courses/courseApiSlice"
+import { useAddOwnerMutation, useAddTutorMutation, useGetOwnersQuery, useGetTutorsQuery, useRemoveOwnerMutation, useRemoveTutorMutation } from "../../../features/courses/courseApiSlice"
 import { useEffect, useState } from "react";
 
-import { selectCurrentUser } from '../features/auth/authSlice';
-import { useGetTeachersQuery } from '../features/users/usersApiSlice';
+import { selectCurrentUser } from '../../../features/auth/authSlice';
+import { useGetTeachersQuery } from '../../../features/users/usersApiSlice';
 import { useSelector } from 'react-redux';
 
 const TeacherList = (props: any) => {
