+++ frontend/src/components/Dashboard/Dashboard.tsx
@@ -2,12 +2,12 @@ import { Button, Container, Dropdown, Form, ListGroup } from 'react-bootstrap'
 import { useEffect, useState } from 'react'
 
 import { CaretDownFill } from 'react-bootstrap-icons'
-import Header from './Header'
+import Header from '../Header/Header'
 import React from 'react'
 import ResultsTable from './ResultsTable'
-import { useGetCoursesSessionsExercisesQuery } from '../features/courses/courseApiSlice'
+import { useGetCoursesSessionsExercisesQuery } from '../../features/courses/courseApiSlice'
 
-const Results = () => {
+const Dashboard = () => {
 
     const [currentSession, setCurrentSession] = useState(-1)
     const [currentSessionTitle, setCurrentSessionTitle] = useState('')
@@ -189,4 +189,4 @@ const Results = () => {
     </>) : (<></>)
 }
 
-export default Results
+export default Dashboard
