+++ frontend/src/components/Header/Header.tsx
@@ -1,12 +1,11 @@
 import { Container, Dropdown, Nav, Navbar } from 'react-bootstrap'
-import { selectCurrentUser, selectIsTeacher } from '../features/auth/authSlice'
+import { selectCurrentUser, selectIsTeacher } from '../../features/auth/authSlice'
 
 import { PersonCircle } from 'react-bootstrap-icons'
 import React from 'react'
-import { logOut } from '../features/auth/authSlice'
+import { logOut } from '../../features/auth/authSlice'
 import { useDispatch } from 'react-redux'
 import { useKeycloak } from '@react-keycloak/web'
-import { useNavigate } from 'react-router-dom'
 import { useSelector } from 'react-redux'
 
 const Header = () => {
@@ -14,11 +13,10 @@ const Header = () => {
     const username = useSelector(selectCurrentUser)
     const isTeacher = useSelector(selectIsTeacher)
     const dispatch = useDispatch()
-    const navigate = useNavigate()
     const { keycloak } = useKeycloak()
 
     const signout = async () => {
-        await dispatch(logOut({}))
+        dispatch(logOut({}))
         keycloak.logout({
             redirectUri: window.location.origin + '/'
         })
@@ -32,7 +30,7 @@ const Header = () => {
                 <Navbar.Collapse id="basic-navbar-nav">
                     <Nav className="me-auto">
                         <Nav.Link href="/course">Courses</Nav.Link>
-                        {isTeacher ? (<Nav.Link href="/results">Results</Nav.Link>) : (<></>)}
+                        {isTeacher ? (<Nav.Link href="/dashboard">Dashboard</Nav.Link>) : (<></>)}
                     </Nav>
 
                     <Dropdown align="end">
