+++ frontend/src/components/Util/CodePreview.tsx
@@ -1,5 +1,5 @@
 import Editor from "@monaco-editor/react";
-import { useGetSubmissionFileContentQuery } from '../features/submission/submissionApiSlice';
+import { useGetSubmissionFileContentQuery } from '../../features/submission/submissionApiSlice';
 
 const CodePreview = (props: any) => {
